Use ClimaUtilities inputs in run_amip
ClimaUtilities comes with a rich array of tools to read input data.
Among these are the `SpaceVaryingInput` and the `TimeVaryingInput` to
read data from NetCDF files (and more). These functions are more
efficient, versatile, and come with various interpolation options.

This commit moves AMIP to using those. In the process, it moves from
using ClimaCoreTempestRemap to Interpolations.