delete Regridder.cgll2latlonz
We have a function `Regridder.cgll2latlonz` that is a wrapper around
wrappers around ClimaCoreTempestRemap functions. We have no use for it
in source, and it is only called in a single test. This commit deletes
the function, and calls the CCTR wrappers directly from the test where
`cgll2latlonz` was called before.