Use try catch directly for artifact
The artifact_exists function didn't really work as expected. This is because the
Artifacts.toml file that was being used is the one for ClimaCoupler (ie, not
existing). The Artifacts.toml file used is the one associated to the macro call,
so the macro has to be next to its Artifacts.toml. So, I just explicitly call
try catch instead of moving it into a function.