Refactor leaderboard code
This commit refactors the leaderboard code using the new features from
ClimaAnalysis. This commit also deletes the old leaderboard code and the
tests for it. Also, there is an off by one month issue when handling the
dates and seasons. This arises because the simulation data assigns the
next month the monthly average for this month. For instance, the monthly
average for January 2010 is assigned the date 2/1/2010. This is fixed in
this commit.

The commit also moves the code to leaderboard.jl and data_sources.jl.
The file read_amip.jl calls leaderboard.jl to plot now. The conditional
to run leaderboard.jl is changed from t_end > 86400 to t_end
> 86400 * 31 * 3 since it doesn't make sense to run the code unless
monthly averages are computed. This is because the dataset contains
monthly averages.

One significant difference is the leaderboard which now plot the best
and worst single model using only annual rather than averaging the error
over annual and seasonal data.