Add specific component dt's to cli_options and parse (#1059)
* Add specific component dt's to cli_options and parse

If both dt and component dts are passed in, the component dts are chosen.
When both are passed, or if the coupler dt is not divisible by a component dt,
a relavant warning or error message is passed.

* Add experiment for component dts to pipeline

* Improve comments and change config_dict logic for dts

Unused dt's are now removed from the config.
IF dt_atmos is used, it is temporarily duplicated
as dt during get_atmos_config_dict

* Fix how get_atmos_config_dict accesses configs

get_atmos_config_dict accessed config_dict from the global
scope instead of coupler_dict, which is passed as an argument

* Add dt through merge rather than modifying coupler_dict