Refactor independent time step argument reads
The code that read the component dts used eval
and symbols to create the component dt vars.

This is changed to keep them in a dict. dt_atmos is
held in this dict, but it is not actually used. I kept
it there for consistency, but it could be removed.