Replace Plots usage with Makie
All uses of Plots and ClimaCorePlots are removed in this commit.
For almost all of the generated plots, the Makie version is
visually similar to the Plots.jl versions.

Due to a limitation of Makie, the sea_breeze output animations
are slightly different. The Plots.jl animations had a colorbar and levels
that rescaled each frame. This commit makes the animations use a
consistent colorbar and levels.

Plots and ClimaCorePlots are removed from Project.toml of both
experiments/ClimaEarth and experiments/ClimaCore.

Some of the contour plots look slightly different due to
differences in Makie and Plots.jl graphics, and because the default
behavior of ClimaCoreMakie is to use viridis.