Remove regridding funcs that are in ClimaUtilities
Functions write_to_hdf5, reshape_cgll_sparse_to_field!,
hdwrite_regridfile_rll_to_cgll, get_coords, get_time, and write_to_hdf5
are removed. The hdwrite_regridfile_rll_to_cgll only works for
2d spaces in ClimaUtilities, and the function in Regridder.jl also
worked for 3d spaced, so this functionality is lost.

The functions read_available_dates and yyyymmmddd_to_datetime
are copied from ClimaUtilities. This is necessary because
land_fraction needs the first date in the dataset. TempestRegridder
calls hdwrite_regridfile_rll_to_cgll, which extracts the dates and
returns them, but TempestRegridder then discards those extracted dates.

The regridder functions in ClimaUtilities do not work properly when
the time dim is made up of numbers, and it errors when the time dim
is made up of large floats. To function properly, the time time must be
datetimes. The functions in Regridder.jl worked with the time dim as numbers.

Many of the regridding functions in Regridder.jl allowed the outfile_root
to be set. When using the TempestRegridder, it is autmatically set to
the varname. This results in file with names structured like:
"varname_varname_time.hdf5"