Simplify callbacks
This commit removes most of the callback infrastructure in
`ClimaCoupler` in favor of a simpler one that leverages what we already
use in other packages. The current "schedule" system is the one from
`ClimaDiagnostics`, but the plan is to move it to `ClimaUtilities` as
part of the work towards the integer time.

Using the schedule system from `ClimaDiagnostics` eliminated the need
for keeping track of the first day of the month.

I also updated the checkpointing callbacks so that they can run with
frequencies that are smaller than one hour.