Tr/isolate run amip postprocessing (#1120)
* Replace mode_name with mode_type in run_amip

This commit adds a struct for each different mode_name.
A struct is returned by get_cupler_args as mode_type.
This is used everywhere mode_name was. The structs do have hierarchy,
but the Abstract types should be given better names.

* Isolate run_amp.jl postprocessing

This commit adds a new file,
experiments/ClimaEarth/user_io/postprocessing.jl, which contains
the `postprocess_sim` function. This function copies the if statements
previously at the end of run_amip.jl, but it uses dispatch for the simulation
mode.

* Improve comments and docs in postprocessing.jl

* Add docstrings for simulation mode types

* Fix NEWS.md formatting

* Add News.md section for simulation mode structs and postprocessing

* Change simulation mode structs to abstract types

This is more in line with the ComponentModelSimulation types
in Interfacer.jl

* Fix documentation errors