Add restart test and support for restarts
This commit implements a way to restart simulations by saving both state
and caches of component models, as well as the coupler fields.

Given that caches are complex object, I implemented this using JLD2
files.

The challenges with JLD2 files are that:
- they are not MPI compatible,
- they are not GPU compatible.

For this reason, I have to move everything to the CPU, and have each
process write to its own output. This adds a restriction: only the same
number of MPI process (and the same machine) can be used for restarts.

In addition to this, this approach requires component models to
implement their functions to restore their caches.

Something that can be improved in the future is that, ClimaAtmos is
currently producing two checkpoints, one independently, and one from
ClimaCoupler. This should not be needed, but it is currently needed
because there's no other way to start ClimaAtmos at a different time.

The other problem here is that the MPI test occasionally hangs (as it
does in ClimaAtmos).