Tr/add get_fields and update_fields to ClimaAtmosSimulation (#1205)
* Add more get_fields to ClimaAtmosSimulation

Add diffuse_fraction, SW_d, LW_d, and cos of the zenith_angle
to get_field on AtmosModelSimulation in Interfacer.jl.

Add get_field methods for the above values in
ClimaAtmosSimulation

Remove safe_div and use anonymous func

Only calculate on first level for diffuse_frac

Add update_field for emissivity in AtmosModelSim

Add emissivity to update_field in the Interfacer,
and add a method definition for it in climaatmos.jl
All 16 bands are set to the same values.

Clamp diffuse_fraction to [0,1]

Remove :co2 from atmos update_field

:co2 is removed from the default AtmosModelSimulation
update_field. The method definition is also removed for
ClimaAtmosSimulation.

Add atmos get_field for :co2

Add a default get_field for :co2 with AtmosModelSimulation,
and a method definition for ClimaAtmosSimulation. Note that this
returns a scalar.

Add new fields to interfacer_tests

Revert to returning zero when dividing by zero

In get_field(::ClimaAtmosSimulation, ::Val{:diffuse_fraction}

* Add air pressure, air temp, and humidity get_field

Add them to the the generic get_field for
AtmosModelSimulations and specific methods
for ClimaAtmosSimulations

* Update Interfacer.md with atmos get_field changes

* Change `cos_zenith_angle` to `cos_zenith`

It is refered to as `cos_zenith` by atmos and
`cos\thetas` by land. This unifies it with one of
the models.
[skip ci]

* Change humidity to specific_humidity for atmos

* Remove not always required fields default get_field

remove from interfacer tests