Fix restarting simulation from command line
This commit fixes a broken option in ClimaCoupler. ClimaCoupler would
not restart correctly when the details for the restart were passed via
command line, unless `restart_dir` was the same as `output_dir_root`.

Version 0.29.1 of ClimaAtmos introduces a way to go around the
underlying problem that introduced that bug and also introduced a second
limitation to the restart system.

Closes #1218