Unify Interfacer argname for simulation (#1225)
The methods for each component in experiments/ClimaEarth/components/
refer to the simulation argument with different names. For the
ClimaAtmosSimulation, it is refered to as both `sim` and `atmos_sim`.
Similarly, the BucketSimulation is refered to as both `sim` and
`bucket_sim`. This commit unifies them by changing everything to `sim`.
The opposite could also be done.