Switch to PartionedStateFluxes
`CombinedStateFluxes`, while simple, is not super physically less
sound (combining states of different component models into a single
average state is necessarily the best idea). In addition to this,
`CombinedStateFluxes` relies on the infrastructure in ClimaAtmos to
compute turbulent fluxes, which is highly complex and difficult to
understand.

This commit fixes `PartionedStateFluxes` and switches all the jobs to
use this system. `PartionedStateFluxes` was broken for certain jobs
because it was not setting certain values, such as ustar, which are
needed (e.g., in EDMF).

In this commit, we also change how surface temperature is propagated to
atmos. Here, we take directly the area-weighted value computed from the
surface models and directly set it in atmos.