Split `setup_and_run` in multiple steps
This commit introduces more structure in ClimaEarth. It splits
`setup_and_run` in three functions:
- `setup`, which produces a `CoupledSimulation` (hence, it was renamed
to be `CoupledSimulation`, since it is just a constructor)
- `run!`, which runs the simulation. This is supported by the
introduction of `step!`, to take one single coupling step. This is
particularly useful in debugging
- `postprocess`, to make plots

The advantage of this separation is that one can now construct a
simulation without running it, and step a simulation one step at a time.
Both of these are particularly useful in interactive work and debugging.

This commit also renames `calendar_dt` to `diag_calendar_dt` to make it
clearer that it refers to diagnostics.