Fix insolation update
Switching to PartitionedStateFluxes introduced a bug where insolation
was not updated. This is because the correct `update_sim!` method was
not being called.

This commit fixes that. In parallel, I also fixed the update of
insolation in ClimaAtmos so that it is automatically updated every time
RRTGMP is called.

https://github.com/CliMA/ClimaAtmos.jl/pull/3788

As I discuss in that issue, I think the only reason for updating the
insolation tuple in ClimaCoupler is to do it more often that the
radiation timestep. I think this is because the `cos_zenith_angle` is
one of the exchanged field. So, if we remove `cos_zenith_angle` from the
exchange fields, we can remove this specialized method in
`climaatmos.jl` and use the default one in `FieldExchanger`.