Remove test bash script job
The "Submit and Monitor sbatch Job" soft-fails all the time and no one
really looks at it. The job also submits its own slurm jobs, which are
not tracked by buildkite. This is problem because when a build is
canceled, those jobs become stray (and are not cleaned up)

This commit removes the job.