Add `get_field` with `remap`
This commit adds a method to `Interfacer.get_field`. This method takes
in a `target_space` and remaps the field onto the space. This is needed
to support Oceananigans fields.

Only the trivial case with `target_space` same as original space is
currently supported.

In the future, we need to:
- Add a non-allocating version
- Carefully handle non-scalars (or just exchange scalars)
- Add support for remapping across different spaces