Change coupling loop end cond
The coupling loop end condition is when the simulation
time is greater than the end time. This changes it
to be when it is greater than or equal to the end time.

This is done to prevent the ClimaLand driver cb from
erroring when dt is a divisor of t_end. It will still error when
dt is not a multiple of dt.