Protect against missing nonorgraphic gravity waves
In ClimaAtmos 0.30, the tendency for nonographic gravity waves was
turned into a callback. With the recent refactoring, this callback has
to be explicitly called during initialization. However, the code crashes
when nonographic gravity waves are not enabled in the atmos model.

I fixed this in https://github.com/CliMA/ClimaAtmos.jl/pull/3792, but
version 0.30 remains buggy. This commit works around that.